clean up mujoco tests
